+++ packages/components/tsup.config.ts
@@ -16,6 +16,10 @@ export default defineConfig(options => ({
         options.assetNames = "assets/fonts/[name]";
     },
     async onSuccess() {
+        // In watch mode we will build using this function.
+        // If not in watch mode, we run tsc separately
+        // to make sure PR validation works
+        if (!options.watch) return;
         // eslint-disable-next-line no-console
         console.log("⚙️ Generating typescript declarations..");
         spawnSync("tsc", ["--project", "tsconfig.json", "--emitDeclarationOnly", "--declaration"]);