+++ packages/core/src/hooks/useCoreStackNavigation.ts
@@ -1,0 +1,4 @@
+import { NavigationProp, useNavigation } from "@react-navigation/native";
+import { CoreStackParamListBase } from "../types";
+
+export const useCoreStackNavigation = () => useNavigation<NavigationProp<CoreStackParamListBase>>();