+++ packages/components/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @equinor/mad-components
 
+## 0.8.0
+
+### Minor Changes
+
+-   fb69a74: Fixed flex issues on search
+
+## 0.7.0
+
+### Minor Changes
+
+-   b8b7a72: changed autocomplete to have sub component `multiselect`. Removed `Multiselect` prop in
+    `autocomplete` component. Use `autocomplete.multiselect` instead.
+
+### Patch Changes
+
+-   c2f7949: fix typescript errors
+-   c2f7949: fix label and iconColor errors in TextField component
+
+## 0.6.9
+
+### Patch Changes
+
+-   a29275a: Change Autocomplete to use TextField instead of Input
+-   26fbbbe: added variant prop for validation handling in input component
+
+## 0.6.8
+
+### Patch Changes
+
+-   e8d96f6: Added styling fixes for Autocomplete component
+
 ## 0.6.7
 
 ### Patch Changes