+++ packages/navigation/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @equinor/mad-navigation
 
+## 0.1.16
+
+### Patch Changes
+
+-   Updated dependencies [fb69a74]
+    -   @equinor/mad-components@0.8.0
+
+## 0.1.15
+
+### Patch Changes
+
+-   Updated dependencies [c2f7949]
+-   Updated dependencies [b8b7a72]
+-   Updated dependencies [c2f7949]
+    -   @equinor/mad-components@0.7.0
+
+## 0.1.14
+
+### Patch Changes
+
+-   Updated dependencies [a29275a]
+-   Updated dependencies [26fbbbe]
+    -   @equinor/mad-components@0.6.9
+
+## 0.1.13
+
+### Patch Changes
+
+-   Updated dependencies [e8d96f6]
+    -   @equinor/mad-components@0.6.8
+
 ## 0.1.12
 
 ### Patch Changes