+++ packages/skia-draw/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @equinor/react-native-skia-draw
 
+## 0.2.26
+
+### Patch Changes
+
+-   Updated dependencies [fb69a74]
+    -   @equinor/mad-components@0.8.0
+
+## 0.2.25
+
+### Patch Changes
+
+-   Updated dependencies [c2f7949]
+-   Updated dependencies [b8b7a72]
+-   Updated dependencies [c2f7949]
+    -   @equinor/mad-components@0.7.0
+
+## 0.2.24
+
+### Patch Changes
+
+-   Updated dependencies [a29275a]
+-   Updated dependencies [26fbbbe]
+    -   @equinor/mad-components@0.6.9
+
+## 0.2.23
+
+### Patch Changes
+
+-   Updated dependencies [e8d96f6]
+    -   @equinor/mad-components@0.6.8
+
 ## 0.2.22
 
 ### Patch Changes